+++ laktory/_settings.py
@@ -1,4 +1,6 @@
+from pydantic import model_validator
 from typing import Union
+from typing import Any
 from settus import BaseSettings
 from settus import Field
 from settus import SettingsConfigDict
@@ -21,8 +23,9 @@ class Settings(BaseSettings):
     aws_secret_access_key: Union[str, None] = Field(None, alias="AWS_SECRET_ACCESS_KEY")
 
     # Databricks
-    workspace_root: str = Field("/.laktory/", alias="LAKTORY_WORKSPACE_ROOT")
-    landing_mount_path: str = Field("/mnt/landing/", alias="LAKTORY_LANDING_MOUNT_PATH")
+    workspace_env: str = Field("dev", alias="LAKTORY_WORKSPACE_ENV")
+    workspace_laktory_root: str = Field("/.laktory/", alias="LAKTORY_WORKSPACE_LAKTORY_ROOT")  # TODO: Review variable name
+    workspace_landing_root: str = Field("", alias="LAKTORY_WORKSPACE_LANDING_ROOT")
 
     # Logging
     log_level: str = Field("INFO", alias="LAKTORY_LOG_LEVEL")
@@ -36,6 +39,14 @@ class Settings(BaseSettings):
     databricks_token: Union[str, None] = Field(None)
     databricks_account_id: Union[str, None] = Field(None)
     databricks_warehouse_id: Union[str, None] = Field(None)
+
+    @model_validator(mode="after")
+    def excluded_fields(self) -> Any:
+
+        if self.workspace_landing_root is "":
+            self.workspace_landing_root = f"/Volumes/{self.workspace_env}/sources/landing/"
+
+        return self
 
 
 settings = Settings()