+++ tests/test_datasources.py
@@ -6,7 +6,7 @@ def test_event_data_source():
         name="stock_price",
         producer={"name": "yahoo_finance"},
     )
-    assert reader.dirpath == "/mnt/landing/events/yahoo_finance/stock_price/"
+    assert reader.root_path == "/mnt/landing/events/yahoo_finance/stock_price/"
 
 
 if __name__ == "__main__":